+++ tsdb/db.go
@@ -90,6 +90,7 @@ func DefaultOptions() *Options {
 		EnableOverlappingCompaction: true,
 		EnableSharding:              false,
 		EnableDelayedCompaction:     false,
+		CacheOnlySeriesSymbols:      false,
 		CompactionDelayMaxPercent:   DefaultCompactionDelayMaxPercent,
 		CompactionDelay:             time.Duration(0),
 		PostingsDecoderFactory:      DefaultPostingsDecoderFactory,
@@ -203,6 +204,9 @@ type Options struct {
 	CompactionDelay time.Duration
 	// CompactionDelayMaxPercent is the upper limit for CompactionDelay, specified as a percentage of the head chunk range.
 	CompactionDelayMaxPercent int
+
+	// CacheOnlySeriesSymbols enables caching of TSDB symbols only when adding series, during compaction.
+	CacheOnlySeriesSymbols bool
 
 	// NewCompactorFunc is a function that returns a TSDB compactor.
 	NewCompactorFunc NewCompactorFunc
@@ -430,15 +434,16 @@ var ErrClosed = errors.New("db already closed")
 // Current implementation doesn't support concurrency so
 // all API calls should happen in the same go routine.
 type DBReadOnly struct {
-	logger     *slog.Logger
-	dir        string
-	sandboxDir string
-	closers    []io.Closer
-	closed     chan struct{}
+	logger                 *slog.Logger
+	dir                    string
+	sandboxDir             string
+	CacheOnlySeriesSymbols bool
+	closers                []io.Closer
+	closed                 chan struct{}
 }
 
 // OpenDBReadOnly opens DB in the given directory for read only operations.
-func OpenDBReadOnly(dir, sandboxDirRoot string, l *slog.Logger) (*DBReadOnly, error) {
+func OpenDBReadOnly(dir, sandboxDirRoot string, cacheOnlySeriesSymbols bool, l *slog.Logger) (*DBReadOnly, error) {
 	if _, err := os.Stat(dir); err != nil {
 		return nil, fmt.Errorf("opening the db dir: %w", err)
 	}
@@ -456,10 +461,11 @@ func OpenDBReadOnly(dir, sandboxDirRoot string, l *slog.Logger) (*DBReadOnly, er
 	}
 
 	return &DBReadOnly{
-		logger:     l,
-		dir:        dir,
-		sandboxDir: sandboxDir,
-		closed:     make(chan struct{}),
+		logger:                 l,
+		dir:                    dir,
+		CacheOnlySeriesSymbols: cacheOnlySeriesSymbols,
+		sandboxDir:             sandboxDir,
+		closed:                 make(chan struct{}),
 	}, nil
 }
 
@@ -509,12 +515,16 @@ func (db *DBReadOnly) FlushWAL(dir string) (returnErr error) {
 	mint := head.MinTime()
 	maxt := head.MaxTime()
 	rh := NewRangeHead(head, mint, maxt)
-	compactor, err := NewLeveledCompactor(
+	compactor, err := NewLeveledCompactorWithOptions(
 		context.Background(),
 		nil,
 		db.logger,
 		ExponentialBlockRanges(DefaultOptions().MinBlockDuration, 3, 5),
-		chunkenc.NewPool(), nil,
+		chunkenc.NewPool(),
+		LeveledCompactorOptions{
+			EnableOverlappingCompaction: true,
+			CacheOnlySeriesSymbols:      db.CacheOnlySeriesSymbols,
+		},
 	)
 	if err != nil {
 		return fmt.Errorf("create leveled compactor: %w", err)
@@ -906,6 +916,7 @@ func open(dir string, l *slog.Logger, r prometheus.Registerer, opts *Options, rn
 		db.compactor, err = NewLeveledCompactorWithOptions(ctx, r, l, rngs, db.chunkPool, LeveledCompactorOptions{
 			MaxBlockChunkSegmentSize:    opts.MaxBlockChunkSegmentSize,
 			EnableOverlappingCompaction: opts.EnableOverlappingCompaction,
+			CacheOnlySeriesSymbols:      opts.CacheOnlySeriesSymbols,
 			PD:                          opts.PostingsDecoderFactory,
 			UseUncachedIO:               opts.UseUncachedIO,
 		})